+++ docs/resources/organization_policy.md
@@ -1,19 +1,19 @@
 ---
 layout: "spotinst"
-page_title: "Spotinst: health_check"
-subcategory: "Elastigroup"
+page_title: "Spotinst: organization_policy"
+subcategory: "Organization"
 description: |-
-  Provides a Spotinst Health Check resource.
+  Provides a Spotinst access policy.
 ---
 
-# spotinst\_administration\_org\_policy
+# spotinst\_organization\_policy
 
-Provides a Spotinst access policy 
+Provides a Spotinst access policy. 
 
 ## Example Usage
 
 ```hcl 
-resource "spotinst_administration_org_policy" "terraform_policy" {
+resource "spotinst_organization_policy" "terraform_policy" {
   name = "test-policy"
   description = "policy by terraform"
   policy_content{
@@ -38,8 +38,8 @@ The following arguments are supported:
 * `name` - (Required) Name of the Policy.
 * `description` - Short description of policy.
 * `policy_content` - Set permissions objects list.
-* `statements` - List of permissions statements.
-    * `actions` - Set a list of required actions for this permissions statement.
+  * `statements` - List of permissions statements.
+    * `actions` - Set a list of required actions for this permissions statement. Full list of actions can be found in [https://docs.spot.io/account-user-management/user-management/access-policies-actions/](https://docs.spot.io/account-user-management/user-management/access-policies-actions/)
     * `effect` - Valid values "ALLOW", "DENY".
     * `resources` - Set a list of resources IDs.
