+++ spotinst/organization_policy/fields_spotinst_organization_policy.go
@@ -1,4 +1,4 @@
-package administration_org_policy
+package organization_policy
 
 import (
 	"fmt"
