+++ spotinst/organization_user_group/fields_spotinst_organization_user_group.go
@@ -1,4 +1,4 @@
-package administration_org_user_group
+package organization_user_group
 
 import (
 	"fmt"
