+++ spotinst/organziation_programmatic_user/fields_spotinst_organization_programmatic_user.go
@@ -1,4 +1,4 @@
-package administration_org_programmatic_user
+package organziation_programmatic_user
 
 import (
 	"fmt"
